+++ simple_calc/Utils_NjoySpectrumPlotter.py
